+++ renderer/src/services/interfaces/CategoryGroupService.ts
